fix: add type assertions for viem readContract responses
- Add 'as bigint' type assertion to all balance checks
- Fix TypeScript errors in forum/posts, forum/threads, and token/balance routes
- Build now passes successfully with no type errors

+++ app/api/forum/posts/route.ts
@@ -93,7 +93,7 @@ export async function POST(req: NextRequest) {
         abi: TOKEN_CONFIG.abi,
         functionName: 'balanceOf',
         args: [wallet as Address],
-      });
+      }) as bigint;
 
       const minRequired = BigInt(TOKEN_CONFIG.requirements.createPost);
